fill main_example.ipynb, move cimsolver to BinaryProblem
