--- SimCIM/M-LOOP_archives/controller_archive_2019-08-16_17-09.txt
@@ -1,20 +1,0 @@
-archive_type='controller'
-num_out_params=46
-out_params=[array([ 2.        , -4.        , -0.04575749,  0.17609126, -0.52287875,       -1.        ]), array([-0.08177653, -4.82187236, -1.01068037,  1.30096353,  1.64144631,        2.73346401]), array([-0.55893736, -3.63167555, -1.09470131,  0.2344187 ,  3.1790302 ,       -0.21326496]), array([ 4.72409848, -5.21877595, -1.9398742 , -1.24102684,  2.33733079,        2.67189863]), array([ 4.94760759, -4.67649885, -1.29830506,  0.33540204, -3.4567469 ,        0.2856665 ]), array([ 2.69457359, -5.87689033, -1.77671735,  1.05634369, -0.74140623,       -0.14728693]), array([ 0.52697373, -5.71317317, -0.31438734,  1.58582404, -3.30593349,       -2.22180902]), array([-1.0420966 , -1.28787593, -1.6484044 ,  0.81068025, -1.08575188,       -3.94678402]), array([-1.18545843, -3.52825852, -1.95971333,  1.57110881, -4.57213119,       -3.36757885]), array([ 1.9551506 , -1.26279105, -1.15333936,  1.2336848 , -4.57087138,        1.29735274]), array([-1.09444786, -5.63564391, -1.49888453,  0.0192893 ,  1.65550067,        1.31338821]), array([-0.19685144, -1.63305132, -2.1077876 , -0.32711925, -4.43605229,       -2.4277048 ]), array([ 4.69435557, -3.72332588, -1.48680312,  1.77362549, -2.90987308,        2.63132239]), array([ 5.        , -7.        ,  0.        , -1.82390874, -5.        ,       -4.        ]), array([-1.5       , -7.        , -0.67395686, -1.60826763, -5.        ,       -3.75606337]), array([ 4.46375673, -1.        , -2.54575749,  0.92689009, -4.68620655,       -4.        ]), array([ 1.83389246, -5.97015325, -1.49396614,  0.08989898,  2.19373785,       -2.90075615]), array([-1.5       , -1.        , -2.54575749, -0.48476307,  4.47712125,       -4.        ]), array([-0.75364964, -7.        , -1.6853063 ,  2.        ,  4.47712125,       -2.89231904]), array([ 5.        , -6.26334345, -2.37041733, -1.82390874, -4.00182253,        3.        ]), array([ 4.14739802, -5.60583694, -2.52603191,  0.02948148, -4.29514835,        1.53960804]), array([ 5.        , -3.53338877, -0.41352666, -1.82390874, -5.        ,        2.01139363]), array([ 5.        , -1.        ,  0.        ,  2.        ,  4.47712125,        0.26823455]), array([-1.5       , -2.93823197, -2.54575749,  0.47499648, -3.47540559,        3.        ]), array([ 1.86772976, -1.95544031, -0.16463342,  1.85811349, -0.26307838,       -2.6982383 ]), array([ 2.44061011, -1.31430782,  0.        , -1.82390874,  4.47712125,       -4.        ]), array([-0.50730349, -1.        ,  0.        ,  2.        , -0.60701242,        2.15012853]), array([-0.73987351, -1.        , -0.13050239, -1.82390874, -5.        ,        3.        ]), array([ 0.20297804, -5.55235964, -1.27915563,  1.27732771, -1.91728697,       -3.87565419]), array([ 5.        , -4.62442071, -0.18135606,  1.89190175,  2.93694769,       -4.        ]), array([ 2.71292824, -1.        , -1.85414498, -1.82390874, -5.        ,        1.96073321]), array([ 0.8477586 , -3.74646072, -2.54575749, -0.07701657, -5.        ,       -4.        ]), array([-0.62842754, -4.98241781, -2.1745431 ,  1.15352258, -1.14819548,       -0.75118168]), array([ 5.        , -1.        ,  0.        , -1.56825746,  4.36749162,        3.        ]), array([ 1.35716793, -7.        , -2.03861826, -0.52429962, -5.        ,       -4.        ]), array([-0.6169778 , -2.90579985, -2.39983906,  0.86179091, -1.15360309,       -1.57649401]), array([ 0.58285943, -6.3800705 , -1.17009475,  0.1533507 ,  2.87888223,       -0.28736072]), array([ 5.        , -4.75002747, -2.54575749,  2.        ,  4.47712125,        0.50952123]), array([ 4.66776048, -3.0058434 , -1.66323243, -0.19171178, -4.51140819,       -2.63860575]), array([ 5.        , -7.        , -1.35921773, -1.36146434, -1.01146591,        0.8157434 ]), array([-0.81662762, -6.82022218, -1.94054163,  0.19671217, -4.06205804,        2.96614019]), array([ 2.97319952, -3.40141294, -2.54575749, -1.22457801,  1.50082212,       -4.        ]), array([ 4.09429064, -5.09187564, -0.22533761,  0.67950111,  2.05739489,        3.        ]), array([ 3.70265441, -6.76007594, -2.54575749,  0.10410475,  0.75628137,        3.        ]), array([ 0.01915393, -6.35136437, -0.36191206,  0.47250903, -3.62141942,       -1.27739351]), array([ 5.        , -2.95683668,  0.        , -1.82390874, -5.        ,       -3.69915026])]
-out_type=['differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'neural_net', 'neural_net', 'neural_net', 'differential_evolution', 'neural_net', 'neural_net', 'neural_net', 'differential_evolution', 'neural_net', 'neural_net', 'neural_net', 'differential_evolution', 'neural_net', 'neural_net', 'neural_net', 'differential_evolution', 'neural_net', 'neural_net', 'neural_net', 'differential_evolution', 'neural_net', 'neural_net', 'neural_net', 'differential_evolution', 'neural_net', 'neural_net', 'neural_net', 'differential_evolution', 'neural_net', 'neural_net', 'neural_net', 'differential_evolution', 'neural_net']
-out_extras=[{}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}]
-in_costs=[0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]
-in_uncers=[0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]
-in_bads=[False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False]
-in_extras=[{}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}]
-max_num_runs=30000.0
-start_datetime='2019-08-16_17-09'
-training_type='differential_evolution'
-num_params=6
-min_boundary=array([-2., -7., -5., -2., -5., -4.])
-max_boundary=array([ 5., -1.,  0.,  2.,  5.,  4.])
-num_in_costs=45
-best_cost=0.0
-best_uncer=0.0
-best_params=array([ 2.        , -4.        , -0.04575749,  0.17609126, -0.52287875,       -1.        ])
-best_index=1
