--- SimCIM/M-LOOP_archives/controller_archive_2019-08-16_17-32.txt
@@ -1,20 +1,0 @@
-archive_type='controller'
-num_out_params=16
-out_params=[array([ 2.        , -4.        , -0.04575749,  0.17609126, -0.52287875,       -1.        ]), array([ 4.53668732, -4.77699106, -1.81651834,  0.18622034,  2.61046735,       -1.97652946]), array([ 4.3981346 , -2.21369967, -1.51954498,  0.45697666, -2.33007683,        1.08331501]), array([ 1.42016059, -4.56478418, -1.34954425, -1.16747509, -1.18759548,        3.52684928]), array([ 3.91834434, -3.84808843, -3.67523023,  0.75010587, -4.31067414,        1.41408282]), array([ 1.2999916 , -4.73929023, -3.11087449,  0.48674562, -4.57351171,        0.57482354]), array([ 4.83017819, -3.09501082, -0.78363315,  0.12016359,  1.97343764,       -0.533577  ]), array([ 1.20826435, -4.35222469, -1.3811507 , -0.3999378 ,  0.74616224,        2.93566863]), array([ 3.2251578 , -3.95255676, -3.00296243, -0.14434103, -1.69628477,        0.75855022]), array([ 2.94046236, -2.68117295, -3.74428143,  0.90914565, -3.17261488,        0.59359211]), array([ 1.28695556, -2.51695011, -1.99106003, -0.38533853,  2.152983  ,       -0.93500008]), array([ 1.1429961 , -2.8521169 , -2.08095137,  1.88114119,  0.62533018,        1.88971677]), array([ 3.21111551, -2.85237039, -0.66400138, -1.13563235,  1.76027517,        3.95950515]), array([ 4.3981346 , -2.21369967, -1.51954498,  0.45697666, -2.33007683,        1.08331501]), array([ 4.3981346 , -2.21369967, -1.51954498,  0.45697666, -2.33007683,        1.08331501]), array([ 4.3981346 , -2.21369967, -1.51954498,  0.45697666, -2.33007683,        1.08331501])]
-out_type=['differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'neural_net', 'neural_net', 'neural_net']
-out_extras=[{}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}]
-in_costs=[0.0, 0.0, -1.7475354496809814e-06, 0.0, 0.0, 0.0, -1.7475354496809814e-06, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, -1.7475354496809814e-06, -1.7475354496809814e-06]
-in_uncers=[0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]
-in_bads=[False, False, False, False, False, False, False, False, False, False, False, False, False, False, False]
-in_extras=[{}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}]
-max_num_runs=30000.0
-start_datetime='2019-08-16_17-32'
-training_type='differential_evolution'
-num_params=6
-min_boundary=array([-2., -7., -5., -2., -5., -4.])
-max_boundary=array([ 5., -1.,  0.,  2.,  5.,  4.])
-num_in_costs=15
-best_cost=-1.7475354496809814e-06
-best_uncer=0.0
-best_params=array([ 4.3981346 , -2.21369967, -1.51954498,  0.45697666, -2.33007683,        1.08331501])
-best_index=3
