--- SimCIM/M-LOOP_archives/controller_archive_2019-08-16_17-50.txt
@@ -1,20 +1,0 @@
-archive_type='controller'
-num_out_params=45
-out_params=[array([ 2.        , -4.        , -0.04575749,  0.17609126, -0.52287875,       -1.        ]), array([ 1.50296988, -1.13187474, -2.09102431,  1.30116991,  3.17047857,       -1.06810398]), array([ 2.41636517, -6.68962063, -1.14482656, -0.02940571, -2.11241363,       -1.12968746]), array([ 0.67521313, -1.79730176, -1.43124558, -1.06222883, -2.17548722,        2.84846222]), array([ 4.3898084 , -5.16177975, -2.2081003 , -0.75666775,  2.98054642,       -3.29313389]), array([ 0.67979547, -5.27532817, -0.5556885 , -0.91822809,  2.65346   ,       -1.57313677]), array([ 1.18660471, -2.14819454, -0.21624044, -1.29511559, -3.87638285,        1.47544951]), array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129]), array([ 3.49649534, -6.5515217 , -2.40227594, -1.48863435, -0.19375973,       -1.51008598]), array([ 2.97475362, -3.33802213, -0.08289632, -0.35321986,  0.5571384 ,       -3.68140154]), array([ 2.9547989 , -4.13171443, -2.83289061,  0.17277502, -1.13191692,       -1.85483487]), array([ 1.24478112, -2.17122549, -2.70600808,  0.14893091,  4.79255293,        1.69342344]), array([ 4.96171981, -5.55518532, -0.03580444, -0.1628218 , -1.15132338,        0.20401847]), array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129]), array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129]), array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129]), array([-0.51804506, -4.5305864 , -2.4226473 , -1.88640234, -1.84953873,       -3.24129476]), array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129]), array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129]), array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129]), array([-0.27080942, -2.2064409 , -0.25106634,  0.64191603,  3.25481016,        0.58172484]), array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129]), array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129]), array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129]), array([-0.48610264, -1.07349522, -1.71070597,  0.32269352, -2.65828817,       -2.62845034]), array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129]), array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129]), array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129]), array([ 4.47906233, -2.11523475, -2.86633246, -0.0814996 ,  2.63059407,        1.27912411]), array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129]), array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129]), array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129]), array([-0.72260281, -3.71372093, -2.38199746, -1.6909303 ,  3.52150236,       -2.77311828]), array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129]), array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129]), array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129]), array([ 0.12704937, -1.85645028, -2.03726677,  0.66054561, -0.96443003,       -2.23970459]), array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129]), array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129]), array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129]), array([-0.54124598, -5.20421417, -1.28133019, -1.00329786, -0.85377383,       -2.98767586]), array([ 3.69478767, -2.9334226 , -0.05663246, -1.15610287,  1.15378187,       -2.8607352 ]), array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129]), array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129]), array([ 1.02261211, -2.08560756, -3.00342159,  0.26749944, -0.56003507,       -3.43492392])]
-out_type=['differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'differential_evolution', 'neural_net', 'neural_net', 'neural_net', 'differential_evolution', 'neural_net', 'neural_net', 'neural_net', 'differential_evolution', 'neural_net', 'neural_net', 'neural_net', 'differential_evolution', 'neural_net', 'neural_net', 'neural_net', 'differential_evolution', 'neural_net', 'neural_net', 'neural_net', 'differential_evolution', 'neural_net', 'neural_net', 'neural_net', 'differential_evolution', 'neural_net', 'neural_net', 'neural_net', 'differential_evolution', 'neural_net', 'neural_net', 'neural_net', 'differential_evolution']
-out_extras=[{}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}]
-in_costs=[0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, -1.7475354496809814e-06, 0.0, -1.7475354496809814e-06, 0.0, 0.0, 0.0, -1.7475354496809814e-06, -1.7475354496809814e-06, -1.7475354496809814e-06, 0.0, -1.7475354496809814e-06, -1.7475354496809814e-06, -1.7475354496809814e-06, 0.0, -1.7475354496809814e-06, -1.7475354496809814e-06, -1.7475354496809814e-06, 0.0, -1.7475354496809814e-06, -1.7475354496809814e-06, -1.7475354496809814e-06, 0.0, -1.7475354496809814e-06, -1.7475354496809814e-06, -1.7475354496809814e-06, 0.0, -1.7475354496809814e-06, -1.7475354496809814e-06, -1.7475354496809814e-06, 0.0, -1.7475354496809814e-06, -1.7475354496809814e-06, 0.0, 0.0, 0.0, -1.7475354496809814e-06, -1.7475354496809814e-06]
-in_uncers=[0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]
-in_bads=[False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False, False]
-in_extras=[{}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}, {}]
-max_num_runs=30000.0
-start_datetime='2019-08-16_17-50'
-training_type='differential_evolution'
-num_params=6
-min_boundary=array([-2., -7., -5., -2., -5., -4.])
-max_boundary=array([ 5., -1.,  0.,  2.,  5.,  4.])
-num_in_costs=44
-best_cost=-1.7475354496809814e-06
-best_uncer=0.0
-best_params=array([ 2.47376576, -4.02867975, -0.53710078, -1.27603727,  0.95349711,       -2.26444129])
-best_index=8
